Make authorizeGetIdentities, authorizeSignAndPost non-optional
